Building code to extract the tweets that we want from the large scraping files. We are going to reduce the tweets to their id, the tweeter id, the set of hashtags for the tweet, the geo-tag (if it exists), whether or not it is a retweet, and the retweet count. This is represented by the class tweet_tags, and these will be stored in a separate file for each day of tweeting, as a dictionary keyed by tweet id. We will also keep track of a set of all of the tweet ids for all days, so that we can do any duplicate elimination and keep track of the number of tweets overall. We will keep this in a separate file. We will count the hashtag tweets separately from the candidate mention and candidate name files.
    While building this, we will keep track of all of the tweeters who have had a tweet captured by the system. This
is represented by the class tweeter, which contains the user id, the count of followers for that user, the count of
friends for that user, and the number of tweets associated with the user. A dictionary of all of these objects, keyed
by tweeter id, will be written out into a separate file.

Note that this code may work, but doesn't scale-- the time it takes to pickle
the resulting data structures is prohibitive.